pyproject.toml
[tool.poetry]
name = "athame"
version = "0.1.0"
description = "A tool for scheduling allowed/forbidden blocks of execution time for daemon processes."
authors = ["Graham Drakeley <gt.drakeley@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.8"
pendulum = "^2.1.2"
pydantic = "^1.8.2"
py-buzz = "^2.1.3"

[tool.poetry.dev-dependencies]
ipython = "^7.24.1"
ipdb = "^0.13.9"
pylint = "^2.8.3"
rope = "^0.19.0"
pytest = "^6.2.4"
pytest-random-order = "^1.0.4"
pytest-sugar = "^0.9.4"
black = "^21.6b0"
isort = "^5.9.1"
pytest-cov = "^2.12.1"
pre-commit = "^2.15.0"

[tool.black]
line-length = 120

[tool.isort]
src_paths = ["src", "tests"]
multi_line_output = 3
include_trailing_comma = true

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "--random-order"
testpaths = ["tests"]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
